backend/requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.8.1
asn1crypto==1.4.0
attrs==21.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.8.3
cached-property==1.5.2
cbor2==5.4.2
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
docopt==0.6.2
docutils==0.17.1
drf-nested-routers==0.94.1
duplicity==0.8.21
et-xmlfile==1.0.1
fasteners==0.14.1
freezegun==1.0.0
future==0.18.2
GeoIP==1.3.2
geoip2==2.9.0
gevent==21.8.0
greenlet==1.1.2
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
isodate==0.6.1
jdcal==1.0
jeepney==0.7.1
Jinja2==3.0.3
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libsass==0.20.1
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
maxminddb==2.0.3
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
num2words==0.5.10
oauthlib==3.2.0
ofxparse==0.21
olefile==0.46
openpyxl==3.0.9
paramiko==2.9.3
passlib==1.7.4
pexpect==4.8.0
Pillow==9.0.1
polib==1.1.1
protobuf==3.12.4
psutil==5.9.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-stdnum==1.17
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
qrcode==7.3.1
reportlab==3.6.8
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==0.9.1
rjsmin==1.1.0
roman==3.3
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.5.3
systemd-python==234
typing_extensions==4.13.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
vobject==0.9.6.1
wadllib==1.3.6
Werkzeug==2.0.2
xdg==5
xkit==0.0.0
xlrd==1.2.0
XlsxWriter==3.0.2
zeep==4.1.0
zipp==1.0.0
zope.event==4.4
zope.interface==5.4.0
